Update foreach mediator test cases
