+++ integration/mediation-tests/tests-mediator-1/src/test/resources/artifacts/ESB/server/repository/deployment/server/synapse-configs/default/proxy-services/foreachLargeMessageTestProxy.xml
@@ -1,34 +1,36 @@
-<proxy xmlns="http://ws.apache.org/ns/synapse" name="foreachLargeMessageTestProxy" transports="https http" startOnLoad="true" trace="disable">
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+  ~
+  ~ WSO2 Inc. licenses this file to you under the Apache License,
+  ~ Version 2.0 (the "License"); you may not use this file except
+  ~ in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing,
+  ~ software distributed under the License is distributed on an
+  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+  ~ KIND, either express or implied. See the License for the
+  ~ specific language governing permissions and limitations
+  ~ under the License.
+  ~
+  -->
+<proxy xmlns="http://ws.apache.org/ns/synapse" name="foreachLargeMessageTestProxy" transports="https http"
+       startOnLoad="true" trace="disable">
     <description/>
     <target>
         <inSequence>
-            <foreach expression="//ns:CheckPriceRequest" xmlns:ns="http://services.samples">
+            <foreach expression="//foreach">
                 <sequence>
-                    <payloadFactory media-type="xml">
-                        <format>
-                            <ns:getQuote>
-                                <ns:request>
-                                    <ns:symbol>$1</ns:symbol>
-                                </ns:request>
-                            </ns:getQuote>
-                        </format>
-                        <args>
-                            <arg evaluator="xml" expression="//ns:CheckPriceRequest/ns:Code"/>
-                        </args>
-                    </payloadFactory>
-                    <log level="full">
-                        <property name="foreach" value="in"/>
+                    <log level="custom">
+                        <property name="ForEachMediatorTestCase" value="in"/>
                     </log>
                 </sequence>
             </foreach>
-            <send>
-                <endpoint>
-                    <address uri="http://localhost:9000/services/SimpleStockQuoteService"/>
-                </endpoint>
-            </send>
+            <drop/>
         </inSequence>
-        <outSequence>
-            <send/>
-        </outSequence>
+        <outSequence/>
     </target>
 </proxy>
